app/requirements.txt
pandas==1.2.2
requests==2.31.0
streamlit==1.23.1
pyaml==23.12.0
datarobot==3.3.2
python-dotenv==1.0.1